# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MediTrack Pro API — PULSO IoT Pill Dispensing System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Patients ───────────────────────────────────────────────
...
  /patients/{patientId}/adherence:
    get:
      operationId: listAdherenceLogs
      tags: [adherence]
      summary: List dose events for a patient
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dose log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DoseEvent"
  /patients/{patientId}/adherence/{doseEventId}/confirm:
    post:
      operationId: confirmDose
      tags: [adherence]
      summary: Confirm a dose was taken
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
        - name: doseEventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoseConfirmInput"
      responses:
        "200":
          description: Confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoseEvent"
  /patients/{patientId}/adherence/summary:
    get:
      operationId: getAdherenceSummary
      tags: [adherence]
      summary: Get adherence stats and streak
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Adherence summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdherenceSummary"
  /patients/{patientId}/adherence/weekly:
    get:
      operationId: getWeeklyAdherence
      tags: [adherence]
      summary: Get weekly adherence chart data
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Weekly chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeeklyAdherencePoint"
  /patients/{patientId}/adherence/trends:
    get:
      operationId: getAdherenceTrends
      tags: [adherence]
      summary: Get trend flags (e.g. weekend-clustered misses)
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trend analysis
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendFlag"
  /patients/{patientId}/adherence/export:
    get:
      operationId: exportAdherenceReport
      tags: [adherence]
      summary: Export adherence report data
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Export data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportReport"

  # ── Dashboard (wow endpoint) ───────────────────────────────
...
  /patients/{patientId}/notifications:
    get:
      operationId: getNotificationSettings
      tags: [notifications]
      summary: Get notification settings for a patient
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSettings"
    patch:
      operationId: updateNotificationSettings
      tags: [notifications]
      summary: Update notification settings
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationSettingsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSettings"

# ── Schemas ────────────────────────────────────────────────────────────────────
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # Auth
    RegistrationInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [caregiver, patient, pwd]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    ForgotPasswordInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
    AuthResult:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [caregiver, patient, pwd]
        createdAt:
          type: string
          format: date-time

    # Patient
...
    DoseEvent:
      type: object
      required: [id, patientId, medicationId, scheduledAt, status]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        medicationId:
          type: integer
        medicationName:
          type: string
        scheduledAt:
          type: string
          format: date-time
        confirmedAt:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          enum: [pending, taken, missed, snoozed]
    DoseConfirmInput:
      type: object
      required: [confirmedAt]
      properties:
        confirmedAt:
          type: string
          format: date-time
        method:
          type: string
          enum: [tap, voice, auto]
          default: tap
    AdherenceSummary:
      type: object
      required: [patientId, overallRate, currentStreak, takenCount, missedCount, totalCount]
      properties:
        patientId:
          type: integer
        overallRate:
          type: number
          description: Percentage 0-100
        currentStreak:
          type: integer
          description: Consecutive days with all doses taken
        takenCount:
          type: integer
        missedCount:
          type: integer
        totalCount:
          type: integer
        lastUpdated:
          type: string
          format: date-time
    WeeklyAdherencePoint:
      type: object
      required: [date, taken, missed, total]
      properties:
        date:
          type: string
          format: date
        taken:
          type: integer
        missed:
          type: integer
        total:
          type: integer
        rate:
          type: number
    TrendFlag:
      type: object
      required: [id, type, message, severity]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [weekend_misses, evening_misses, low_stock, streak_risk, perfect_week]
        message:
          type: string
        severity:
          type: string
          enum: [info, warning, critical]
        detectedAt:
          type: string
          format: date-time
    ExportReport:
      type: object
      required: [patientId, generatedAt, doseEvents, summary]
      properties:
        patientId:
          type: integer
        patientName:
          type: string
        generatedAt:
          type: string
          format: date-time
        shareToken:
          type: ["string", "null"]
        summary:
          $ref: "#/components/schemas/AdherenceSummary"
        doseEvents:
          type: array
          items:
            $ref: "#/components/schemas/DoseEvent"

    # Dashboard
...
    NotificationSettings:
      type: object
      required: [patientId, pushEnabled, smsEnabled, smsPhone, weeklyDigestEnabled, reminderMinutesBefore]
      properties:
        patientId:
          type: integer
        pushEnabled:
          type: boolean
        smsEnabled:
          type: boolean
        smsPhone:
          type: ["string", "null"]
        weeklyDigestEnabled:
          type: boolean
        reminderMinutesBefore:
          type: integer
        language:
          type: string
          enum: [en, fil]
        largerTextEnabled:
          type: boolean
        highContrastEnabled:
          type: boolean
    NotificationSettingsUpdate:
      type: object
      properties:
        pushEnabled:
          type: boolean
        smsEnabled:
          type: boolean
        smsPhone:
          type: ["string", "null"]
        weeklyDigestEnabled:
          type: boolean
        reminderMinutesBefore:
          type: integer
        language:
          type: string
          enum: [en, fil]
        largerTextEnabled:
          type: boolean
        highContrastEnabled:
          type: boolean