# packages/api-zod/src/generated/types/loginInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MediTrack Pro API — PULSO IoT Pill Dispensing System
 * OpenAPI spec version: 0.1.0
 */

export interface LoginInput {
  email: string;
  password: string;
}