# packages/api-zod/src/generated/types/notificationSettings.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MediTrack Pro API — PULSO IoT Pill Dispensing System
 * OpenAPI spec version: 0.1.0
 */
import type { NotificationSettingsLanguage } from './notificationSettingsLanguage';

export interface NotificationSettings {
  patientId: number;
  pushEnabled: boolean;
  smsEnabled: boolean;
  /** @nullable */
  smsPhone: string | null;
  weeklyDigestEnabled: boolean;
  reminderMinutesBefore: number;
  language?: NotificationSettingsLanguage;
  largerTextEnabled?: boolean;
  highContrastEnabled?: boolean;
}